requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
bcrypt==4.0.1
python-dotenv==1.0.0
ultralytics==8.0.196
opencv-python==4.8.1.78
easyocr==1.7.1
numpy==1.24.3